feature_weather - Organized Location Service - Add NetworkMonitor to maintain network state
Android Studio
- Updated to Eel...forcefully......and had to deal with competitive problems

feature_shopping
- Found an error that hiltViewMode() referencing different view models and this lead to returning empty list.

Next fix:
-none

+++ app/src/main/java/com/example/portfolio/feature_shopping/data/local/entities/SellingItemEntity.kt
@@ -3,10 +3,10 @@ package com.example.portfolio.feature_shopping.data.local.entities
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "selling_items")
+@Entity (tableName = "selling_items")
 data class SellingItemEntity(
     @PrimaryKey(autoGenerate = false)
-    var itemId:Long,
+    var itemId:Int,
     var cartId:Long?,
 
     //for compose test use only
